+++ src/App.vue
@@ -11,10 +11,7 @@
             :tree-data="treeData"
             v-model="ids"
             :options="options"
-            @add-a-child="addAChild"
-            @item-click="itemClick"
-            @item-edit="itemEdit"
-            @item-delete="itemDelete"
+            @handle="handle"
           />
         </div>
         <div class="col-md-6">
@@ -38,21 +35,6 @@
             <div class="checkbox">
               <label>
                 <input type="checkbox" v-model="options.idsWithParent"> 复选是否包含目录
-              </label>
-            </div>
-            <div class="checkbox">
-              <label>
-                <input type="checkbox" v-model="options.showAdd"> 显示添加按钮
-              </label>
-            </div>
-            <div class="checkbox">
-              <label>
-                <input type="checkbox" v-model="options.showEdit"> 显示编辑按钮
-              </label>
-            </div>
-            <div class="checkbox">
-              <label>
-                <input type="checkbox" v-model="options.showDelete"> 显示删除按钮
               </label>
             </div>
           </div>
@@ -79,81 +61,72 @@
     data: function () {
       return {
         ids: [4],
-
         options: {
-          itemName: 'name',
+          label: 'label',
           checkbox: true,
           checkedOpen: true,
           folderBold: true,
-          showAdd: true,
-          showEdit: true,
-          showDelete: true,
-          addClass: 'fa fa-plus-square-o',
-          editClass: 'fa fa-edit',
-          deleteClass: 'fa fa-trash-o',
-          openClass: 'fa fa-angle-right',
-          closeClass: 'fa fa-angle-down',
           idsWithParent: true,
           depthOpen: 0,
-          halfCheckedClass: 'fa fa-minus-square-o fa-fw text-primary',
-          checkedClass: 'fa fa-check-square-o fa-fw text-danger',
-          unCheckedClass: 'fa fa-square-o fa-fw'
+          openIcon: 'fa fa-angle-right',
+          closeIcon: 'fa fa-angle-down',
+          halfCheckedIcon: 'fa fa-minus-square-o fa-fw text-primary',
+          checkedIcon: 'fa fa-check-square-o fa-fw text-danger',
+          uncheckedIcon: 'fa fa-square-o fa-fw'
         },
-
         message: [],
-
         treeData: [{
-          name: '根目录[1]',
+          label: '根目录[1]',
           id: 1,
           children: [
             {
-              name: '一级节点[4]',
+              label: '一级节点[4]',
               id: 4,
               children: [
                 {
-                  name: '二级节点[5]',
+                  label: '二级节点[5]',
                   id: 5,
                   children: [
-                    {name: '三级节点[6]', id: 6},
-                    {name: '三级节点[8]', id: 8},
-                    {name: '三级节点[30]', id: 30},
+                    {label: '三级节点[6]', id: 6},
+                    {label: '三级节点[8]', id: 8},
+                    {label: '三级节点[30]', id: 30},
                     {
-                      name: '三级节点[31]',
+                      label: '三级节点[31]',
                       id: 31,
                       children: [
-                        {name: '四级节点[36]', id: 36},
-                        {name: '四级节点[38]', id: 38},
-                        {name: '四级节点[39]', id: 39},
-                        {name: '四级节点[48]', id: 48}
+                        {label: '四级节点[36]', id: 36},
+                        {label: '四级节点[38]', id: 38},
+                        {label: '四级节点[39]', id: 39},
+                        {label: '四级节点[48]', id: 48}
                       ]
                     }
                   ]
                 },
-                {name: '二级节点[9]', id: 9},
-                {name: '二级节点[10]', id: 10},
+                {label: '二级节点[9]', id: 9},
+                {label: '二级节点[10]', id: 10},
                 {
-                  name: '二级节点[11]',
+                  label: '二级节点[11]',
                   id: 11,
                   children: [
-                    {name: '三级节点[12]', id: 12},
-                    {name: '三级节点[13]', id: 13},
+                    {label: '三级节点[12]', id: 12},
+                    {label: '三级节点[13]', id: 13},
                     {
-                      name: '三级节点[14]',
+                      label: '三级节点[14]',
                       id: 14,
                       children: [
-                        {name: '四级节点[15]', id: 15},
-                        {name: '四级节点[16]', id: 16},
-                        {name: '四级节点[17]', id: 17},
-                        {name: '四级节点[18]', id: 18},
+                        {label: '四级节点[15]', id: 15},
+                        {label: '四级节点[16]', id: 16},
+                        {label: '四级节点[17]', id: 17},
+                        {label: '四级节点[18]', id: 18},
                         {
-                          name: '四级节点[19]',
+                          label: '四级节点[19]',
                           id: 19,
                           children: [
-                            {name: '五级节点[20]', id: 20},
-                            {name: '五级节点[21]', id: 21},
-                            {name: '五级节点[22]', id: 22},
-                            {name: '五级节点[23]', id: 23},
-                            {name: '五级节点[24]', id: 24}
+                            {label: '五级节点[20]', id: 20},
+                            {label: '五级节点[21]', id: 21},
+                            {label: '五级节点[22]', id: 22},
+                            {label: '五级节点[23]', id: 23},
+                            {label: '五级节点[24]', id: 24}
                           ]
                         }
                       ]
@@ -162,35 +135,35 @@
                 }
               ]
             },
-            {name: '一级节点[2]', id: 2},
+            {label: '一级节点[2]', id: 2},
             {
-              name: '一级节点[59]',
+              label: '一级节点[59]',
               id: 59,
               children: [
-                {name: '二级节点[60]', id: 60},
-                {name: '二级节点[61]', id: 61},
-                {name: '二级节点[62]', id: 62},
-                {name: '二级节点[63]', id: 63},
-                {name: '二级节点[64]', id: 64}
+                {label: '二级节点[60]', id: 60},
+                {label: '二级节点[61]', id: 61},
+                {label: '二级节点[62]', id: 62},
+                {label: '二级节点[63]', id: 63},
+                {label: '二级节点[64]', id: 64}
               ]
             },
-            {name: '一级节点[3]', id: 3}
+            {label: '一级节点[3]', id: 3}
           ]
         },
         {
-          name: '根目录[99]',
+          label: '根目录[99]',
           id: 99,
           children: [
-            {name: '二级节点[70]', id: 70},
-            {name: '二级节点[71]', id: 71},
-            {name: '二级节点[72]', id: 72},
-            {name: '二级节点[73]', id: 73},
+            {label: '二级节点[70]', id: 70},
+            {label: '二级节点[71]', id: 71},
+            {label: '二级节点[72]', id: 72},
+            {label: '二级节点[73]', id: 73},
             {
-              name: '二级节点[74]',
+              label: '二级节点[74]',
               id: 74,
               children: [
-                {name: '三级节点[82]', id: 82},
-                {name: '三级节点[83]', id: 83}
+                {label: '三级节点[82]', id: 82},
+                {label: '三级节点[83]', id: 83}
               ]
             }
           ]
@@ -199,20 +172,8 @@
     },
 
     methods: {
-      addAChild (id, item) {
-        this.message.push(`触发添加子节点事件，当前节点${JSON.stringify(item)}`)
-      },
-
-      itemClick (id, item) {
-        this.message.push(`触发点击事件，当前节点${JSON.stringify(item)}`)
-      },
-
-      itemEdit (id, item) {
-        this.message.push(`触发编辑按钮事件，当前节点${JSON.stringify(item)}`)
-      },
-
-      itemDelete (id, item) {
-        this.message.push(`触发删除按钮事件，当前节点${JSON.stringify(item)}`)
+      handle (item) {
+        this.message.push(`节点 ${JSON.stringify(item)}，'handle' 事件`)
       }
     }
   }